+++ Runtime/RyujuEngine.Units/Tempo.cs
@@ -27,39 +27,39 @@ namespace RyujuEngine.Units
 		/// Create an instance with the specified beats per minute.
 		/// BPM でインスタンスを生成します。
 		/// </summary>
-		public static Tempo FromBPM(double beatsPerMinute) => new Tempo(beatsPerMinute);
+		public static Tempo FromBPM(float beatsPerMinute) => new Tempo(beatsPerMinute);
 
-		private Tempo(double beatsPerMinute) => BeatsPerMinute = beatsPerMinute;
+		private Tempo(float beatsPerMinute) => BeatsPerMinute = beatsPerMinute;
 
 		/// <summary>
 		/// The beats per minute.
 		/// 1 分間で刻まれる拍数です。
 		/// </summary>
-		public readonly double BeatsPerMinute;
+		public readonly float BeatsPerMinute;
 
 		/// <summary>
 		/// The beats per second.
 		/// 1 秒間で刻まれる拍数です。
 		/// </summary>
-		public double BeatsPerSecond => BeatsPerMinute / 60.0;
+		public float BeatsPerSecond => BeatsPerMinute / 60.0f;
 
 		/// <summary>
 		/// The seconds per beat.
 		/// 1 拍で刻まれる秒数です。
 		/// </summary>
-		public TimeDuration DurationOfBeat => TimeDuration.OfSeconds(60.0 / BeatsPerMinute);
+		public TimeDuration DurationOfBeat => TimeDuration.OfSeconds(60.0f / BeatsPerMinute);
 
 		/// <summary>
 		/// A flag that indicates a negative tempo.
 		/// テンポが負数かどうかを確かめます。
 		/// </summary>
-		public bool IsNegative => BeatsPerMinute < 0;
+		public bool IsNegative => BeatsPerMinute < 0.0f;
 
 		/// <summary>
 		/// An absolute value of  <see cref="BeatsPerSecond"/>.
 		/// <see cref="BeatsPerSecond"/> の絶対値を返します。
 		/// </summary>
-		public double AbsBeatsPerSecond => IsNegative ? -BeatsPerSecond : BeatsPerSecond;
+		public float AbsBeatsPerSecond => IsNegative ? -BeatsPerSecond : BeatsPerSecond;
 
 		/// <summary>
 		/// An absolute value of <see cref="DurationOfBeat"/>.
@@ -107,15 +107,15 @@ namespace RyujuEngine.Units
 		public static bool operator !=(in Tempo x, in Tempo y) => x.BeatsPerMinute != y.BeatsPerMinute;
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		public static Tempo operator *(Tempo tempo, double scale)
+		public static Tempo operator *(Tempo tempo, float scale)
 			=> new Tempo(tempo.BeatsPerMinute * scale);
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		public static Tempo operator *(double scale, Tempo tempo)
+		public static Tempo operator *(float scale, Tempo tempo)
 			=> new Tempo(tempo.BeatsPerMinute * scale);
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		public static Tempo operator /(Tempo tempo, double divider)
+		public static Tempo operator /(Tempo tempo, float divider)
 			=> new Tempo(tempo.BeatsPerMinute / divider);
 	}
 }
